implement delete user (admin only), create booking role based (Create a new booking with automatic price calculation and vehicle status update), get all booking (admin can sees all, customer sees own)

+++ src/config/db.ts
@@ -35,7 +35,7 @@ export const initDB = async () => {
         vehicle_id INT NOT NULL REFERENCES vehicles(id) ON DELETE CASCADE,
         rent_start_date DATE NOT NULL,
         rent_end_date DATE NOT NULL,
-        total_price NUMERIC(10,2) NOT NULL CHECK (total_price > 0),
+        total_price INT NOT NULL CHECK (total_price > 0),
         status TEXT NOT NULL CHECK (status IN ('active', 'cancelled', 'returned')),
         CONSTRAINT rent_dates_check CHECK (rent_end_date > rent_start_date)
         );
